base-app-02 Build simplest app with docker configuration
