+++ docker-compose.yml
@@ -1,10 +1,7 @@
 services:
-  server:
-    build:
-      context: ./server
-      dockerfile: ./Dockerfile
+  app:
+    build: .
     ports:
       - "3000:3000"
-    volumes:
-      - ./server:/app
-      - .server/rag_qa_llm:/app/rag_qa_llm
+    environment:
+      - NODE_ENV=production
